# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /routes/pois:
    get:
      operationId: getPois
      tags: [routes]
      summary: Points of Interest mit Wikipedia-Anreicherung in einem Kartenausschnitt
      description: >
        Liefert historische und touristische Orte (OpenStreetMap
        historic=*/tourism=attraction|viewpoint) innerhalb einer Bounding Box.
        Orte mit einer verknuepften Wikipedia-/Wikidata-Referenz werden live
        mit einer kurzen Zusammenfassung angereichert (CC BY-SA, mit
        Quellenangabe).
      parameters:
        - name: south
          in: query
          required: true
          schema:
            type: number
        - name: west
          in: query
          required: true
          schema:
            type: number
        - name: north
          in: query
          required: true
          schema:
            type: number
        - name: east
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Points of Interest
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poi"
        "502":
          description: Externe Datenquelle nicht erreichbar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /routes/peaks:
    get:
      operationId: getPeakPois
      tags: [routes]
      summary: Benannte Gipfel in einer Bounding Box
      description: >
        Liefert ausschliesslich benannte OpenStreetMap-Gipfel
        (natural=peak) innerhalb einer Bounding Box. Die Abfrage ist fuer
        das Panorama getrennt vom normalen POI-Download und kann deshalb
        einen groesseren Radius verwenden.
      parameters:
        - name: south
          in: query
          required: true
          schema:
            type: number
        - name: west
          in: query
          required: true
          schema:
            type: number
        - name: north
          in: query
          required: true
          schema:
            type: number
        - name: east
          in: query
          required: true
          schema:
            type: number
        - name: centerLat
          in: query
          required: false
          description: Mittelpunkt der Live-Abfrage; zusammen mit centerLng und radiusKm wird eine Kreisabfrage verwendet.
          schema:
            type: number
        - name: centerLng
          in: query
          required: false
          description: Mittelpunkt der Live-Abfrage.
          schema:
            type: number
        - name: radiusKm
          in: query
          required: false
          description: Radius der Live-Abfrage in Kilometern.
          schema:
            type: number
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Benannte Gipfel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poi"
        "502":
          description: Externe Datenquelle nicht erreichbar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /routes/poi-detail:
    get:
      operationId: getPoiDetail
      tags: [routes]
      summary: Wikipedia/Commons-Anreicherung eines einzelnen POI on demand
      description: >
        Laedt Wikipedia-Zusammenfassung und Bild fuer einen einzelnen Point of
        Interest. Wird erst aufgerufen wenn der Nutzer den POI oeffnet (lazy),
        nicht beim initialen Karten-Laden. Ergebnis wird 24 h serverseitig
        gecacht.
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: kind
          in: query
          required: true
          schema:
            type: string
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: wikipediaTag
          in: query
          required: false
          schema:
            type: string
        - name: wikidataTag
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Anreicherungsergebnis (wiki kann fehlen wenn nichts gefunden)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoiDetailResponse"
        "400":
          description: Ungueltige Anfrage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /routes/custom:
    get:
      operationId: getCustomRoute
      tags: [routes]
      summary: Berechnet eine Wanderroute zwischen zwei selbst gewaehlten Punkten
      description: >
        Berechnet eine Fussweg-Route zwischen Start- und Zielpunkt (OSRM,
        ohne API-Key) und reichert sie mit swisstopo-Hoehenmetern, SAC-Grad
        und Saison-Heuristik an (dieselbe Anreicherung wie bei Kantonsrouten).
        Die Route wird nicht persistiert, sondern bei jeder Anfrage neu
        berechnet.
      parameters:
        - name: startLat
          in: query
          required: true
          schema:
            type: number
        - name: startLng
          in: query
          required: true
          schema:
            type: number
        - name: endLat
          in: query
          required: true
          schema:
            type: number
        - name: endLng
          in: query
          required: true
          schema:
            type: number
        - name: startLabel
          in: query
          required: false
          schema:
            type: string
        - name: endLabel
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Berechnete Route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogRoute"
        "400":
          description: Keine Route zwischen den Punkten gefunden oder Distanz ausserhalb des sinnvollen Bereichs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Externe Datenquelle nicht erreichbar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /routes/custom-waypoints:
    post:
      operationId: planCustomRoute
      tags: [routes]
      summary: Berechnet eine Wanderroute ueber mehrere selbst gesetzte Wegpunkte
      description: >
        Verbindet zwei bis zwoelf geordnete Wegpunkte ueber begehbare Wege
        (Valhalla pedestrian) und reichert die Route mit swisstopo-Hoehenmetern,
        SAC-Grad und Saison-Heuristik an. Die Route wird nicht persistiert.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomWaypointsBody"
      responses:
        "200":
          description: Berechnete Route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogRoute"
        "400":
          description: Ungueltige Wegpunkte, keine Route gefunden oder Distanz ausserhalb des sinnvollen Bereichs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Externe Datenquelle nicht erreichbar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /routes/custom-drawn:
    post:
      operationId: planDrawnRoute
      tags: [routes]
      summary: Mappt eine frei gezeichnete Linie auf begehbare Wege
      description: >
        Nimmt eine vereinfachte, geordnete Freihandlinie entgegen und mappt sie
        mit Valhalla pedestrian auf reale Wanderwege. Die Route wird nicht persistiert.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomDrawnRouteBody"
      responses:
        "200":
          description: Gemappte Route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogRoute"
        "400":
          description: Ungueltige Linie oder kein begehbarer Weg gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Externe Datenquelle nicht erreichbar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /terrain-corridor:
    post:
      operationId: createTerrainCorridor
      tags: [routes]
      summary: Echten SwissTopo-DTM-Korridor entlang einer Route laden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerrainCorridorRequest"
      responses:
        "200":
          description: Routenkorridor mit echten DTM-Höhen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerrainCorridorResponse"
        "400":
          description: Ungültige Routengeometrie
        "502":
          description: Zu wenige verfügbare SwissTopo-Höhenwerte
  /terrain-area:
    post:
      operationId: createTerrainArea
      tags: [routes]
      summary: Rechteckiges SwissTopo-DTM-Gelände um eine vollständige Route laden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerrainAreaRequest"
      responses:
        "200":
          description: Bildschirmfüllende Route-Terrainfläche mit echten DTM-Höhen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerrainAreaResponse"
        "400":
          description: Ungültige Routengeometrie
        "502":
          description: Zu wenige verfügbare SwissTopo-Höhenwerte
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    TerrainCorridorRequest:
      type: object
      required: [geometry]
      properties:
        geometry:
          type: array
          minItems: 2
          maxItems: 500
          items:
            type: array
            prefixItems:
              - type: number
              - type: number
            minItems: 2
            maxItems: 2
        options:
          type: object
          properties:
            rows:
              type: integer
              minimum: 12
              maximum: 80
              default: 32
            columns:
              type: integer
              minimum: 5
              maximum: 13
              default: 9
            halfWidthM:
              type: number
              minimum: 100
              maximum: 1500
              default: 500
    TerrainAreaRequest:
      type: object
      required: [geometry]
      properties:
        geometry:
          type: array
          minItems: 2
          maxItems: 500
          items:
            type: array
            prefixItems:
              - type: number
              - type: number
            minItems: 2
            maxItems: 2
        options:
          type: object
          properties:
            rows:
              type: integer
              minimum: 12
              maximum: 40
              default: 24
            columns:
              type: integer
              minimum: 12
              maximum: 40
              default: 24
            paddingM:
              type: number
              minimum: 500
              maximum: 5000
              default: 2000
            viewportAspect:
              type: number
              minimum: 0.4
              maximum: 1
              default: 0.4615384615
...
    TerrainCorridorResponse:
      type: object
      required:
        [version, source, rows, columns, halfWidthM, routeLengthM, origin, bounds, fetchedAt, grid]
      properties:
        version:
          type: integer
          const: 1
        source:
          type: string
          const: SwissTopo DTM corridor profiles
        rows:
          type: integer
        columns:
          type: integer
        halfWidthM:
          type: number
        routeLengthM:
          type: number
        origin:
          type: object
          required: [lat, lng]
          properties:
            lat:
              type: number
            lng:
              type: number
        bounds:
          type: object
          required: [north, south, east, west]
          properties:
            north:
              type: number
            south:
              type: number
            east:
              type: number
            west:
              type: number
        fetchedAt:
          type: number
        grid:
          type: array
          items:
            type: array
            items:
              $ref: "#/components/schemas/TerrainCorridorCell"
    TerrainAreaResponse:
      type: object
      required:
        [version, source, rows, columns, paddingM, viewportAspect, origin, bounds, fetchedAt, grid]
      properties:
        version:
          type: integer
          const: 1
        source:
          type: string
          const: SwissTopo DTM rectangular route area
        rows:
          type: integer
        columns:
          type: integer
        paddingM:
          type: number
        viewportAspect:
          type: number
        origin:
          type: object
          required: [lat, lng]
          properties:
            lat:
              type: number
            lng:
              type: number
        bounds:
          type: object
          required: [north, south, east, west]
          properties:
            north:
              type: number
            south:
              type: number
            east:
              type: number
            west:
              type: number
        fetchedAt:
          type: number
        grid:
          type: array
          items:
            type: array
            items:
              $ref: "#/components/schemas/TerrainCorridorCell"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CustomRouteWaypoint:
      type: object
      properties:
        lat:
          type: number
          minimum: 45
          maximum: 49
        lng:
          type: number
          minimum: 5
          maximum: 11
      required:
        - lat
        - lng
    CustomWaypointsBody:
      type: object
      properties:
        points:
          type: array
          minItems: 2
          maxItems: 12
          items:
            $ref: "#/components/schemas/CustomRouteWaypoint"
      required:
        - points
    CustomDrawnRouteBody:
      type: object
      properties:
        points:
          type: array
          minItems: 2
          maxItems: 100
          items:
            $ref: "#/components/schemas/CustomRouteWaypoint"
      required:
        - points
...
    CatalogCoordinates:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
      required:
        - lat
        - lng
...
    CatalogRoute:
      type: object
      properties:
        id:
          type: string
        sagaId:
          type: string
        name:
          type: string
        region:
          type: string
        distanceKm:
          type: number
          description: Aus der gespeicherten Geometrie berechnete Streckenlänge in km (weisse Kachel, Navigation).
        distanceTagKm:
          type: number
          description: Amtliche Distanz aus dem OSM-Relation-Tag `distance` (SchweizMobil-Wert); Fallback auf berechnete Geometrie-Distanz wenn kein Tag vorhanden. Immer gesetzt.
        ascentM:
          type: number
        maxElevationM:
          type: number
          description: Hoechster Punkt der Route in Metern ue. M. (swisstopo-Hoehenprofil).
        season:
          type: string
          enum: [ganzjaehrig, eher_sommer, nur_sommer]
          description: >
            Grobe Saison-Einschaetzung aus maximaler Hoehe und SAC-Schwierigkeit
            (Heuristik, keine amtliche Aussage zum aktuellen Zustand).
        minutes:
          type: number
        sac:
          type: string
        sacSource:
          type: string
          nullable: true
          description: Herkunft des SAC-Werts; osm_exact ist ein exakter OSM-Tag, swisstopo_derived eine amtliche Ableitung, unknown unbekannt.
        schweizMobilCondition:
          type: string
          nullable: true
          description: Offizielle SchweizMobil-Kategorie für Kondition (easy, medium oder difficult), nicht auf SAC umgerechnet.
        schweizMobilTechnique:
          type: string
          nullable: true
          description: Offizielle SchweizMobil-Kategorie für Technik (easy, medium oder difficult), nicht auf SAC umgerechnet.
        terrain:
          type: string
        familyFriendly:
          type: boolean
          nullable: true
          description: Konservative technische Familien-Empfehlung aus SAC, Distanz und Aufstieg; null bedeutet unbekannt.
        wheelchairAccessible:
          type: boolean
          nullable: true
          description: Offizielle SchweizMobil-Klassifikation handicap; wird nicht aus Distanz, Höhe oder SAC abgeleitet.
        technicalDifficulty:
          type: string
          nullable: true
        coordinates:
          $ref: "#/components/schemas/CatalogCoordinates"
        geometry:
          type: array
          description: Ausgeduennter Wegverlauf als [lat, lng]-Paare (nur bei realen OSM-Routen vorhanden).
          items:
            type: array
            items:
              type: number
        featured:
          type: boolean
        photoUrl:
          type: string
          nullable: true
          description: Foto-URL aus Wikimedia Commons, bereits in DB gecacht. Null wenn noch kein Foto vorhanden.
        photoAttribution:
          type: string
          nullable: true
          description: Urheber-/Lizenzangabe zum Foto.
        description:
          type: string
          nullable: true
          description: Kurzbeschreibung der Route aus Wikipedia (de); null wenn keine vorhanden.
        descriptionSource:
          type: string
          nullable: true
          description: URL des Wikipedia-Artikels, aus dem die Beschreibung stammt.
        themeKeys:
          type: array
          items:
            type: string
          description: Serverseitig geprüfte Themenbelege der Route.
        qualityStatus:
          type: string
          description: >
            Ergebnis des letzten Plausibilitätschecks: verified, partial, invalid
            oder unverified.
        qualityCheckedAt:
          type: string
          format: date-time
          nullable: true
          description: Zeitpunkt des letzten erfolgreichen Qualitätschecks.
        sources:
          $ref: "#/components/schemas/RouteDataSources"
      required:
        - id
        - sagaId
        - name
        - region
        - distanceKm
        - distanceTagKm
        - ascentM
        - maxElevationM
        - season
        - minutes
        - sac
        - terrain
        - coordinates
        - featured
    RouteDataSource:
      type: object
      properties:
        label:
          type: string
        url:
          type: string
          format: uri
          nullable: true
      required: [label, url]
    RouteDataSources:
      type: object
      properties:
        route:
          $ref: "#/components/schemas/RouteDataSource"
        geometry:
          $ref: "#/components/schemas/RouteDataSource"
        distance:
          $ref: "#/components/schemas/RouteDataSource"
        ascent:
          $ref: "#/components/schemas/RouteDataSource"
        difficulty:
          $ref: "#/components/schemas/RouteDataSource"
      required: [route, geometry, distance, ascent, difficulty]
...
    Poi:
      type: object
      description: >
        Historischer oder touristischer Ort aus OpenStreetMap, optional live mit
        einer Wikipedia-Zusammenfassung angereichert.
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          description: "OSM-Tag, z.B. historic=ruins oder tourism=attraction"
        lat:
          type: number
        lng:
          type: number
        elevation:
          type: number
          nullable: true
          description: "OSM-Höhe in Metern über Meer, sofern am Gipfel gepflegt."
        wiki:
          $ref: "#/components/schemas/WikiSummary"
        wikipediaTag:
          type: string
          nullable: true
          description: "OSM wikipedia-Tag (z.B. 'de:Basiliskenbrunnen Basel'), fuer on-demand-Anreicherung."
        wikidataTag:
          type: string
          nullable: true
          description: "OSM wikidata-Tag (z.B. 'Q123456'), fuer on-demand-Anreicherung."
        osmContext:
          type: string
          nullable: true
          description: "Kuratierter OSM-Kontext (note, inscription, alt_name …) als formatierter String fuer den KI-Prompt."
        source:
          type: string
          description: "Primärquelle des POIs."
        sourceUrl:
          type: string
          format: uri
          description: "Direkter Nachweis des OSM-Objekts."
        checkedAt:
          type: string
          format: date-time
          nullable: true
          description: "Zeitpunkt des erfolgreichen Quellenabrufs."
      required:
        - id
        - name
        - kind
        - lat
        - lng
    PoiDetailResponse:
      type: object
      description: >
        On-demand-Anreicherung eines einzelnen POI mit Wikipedia-Zusammenfassung
        und/oder Bild (kann null sein wenn nichts gefunden wurde).
      properties:
        wiki:
          $ref: "#/components/schemas/WikiSummary"
      required: []
...
    WikiSummary:
      type: object
      description: Live von Wikipedia geladene Kurzzusammenfassung (CC BY-SA).
      properties:
        title:
          type: string
        extract:
          type: string
        url:
          type: string
        lang:
          type: string
        image:
          type: string
          nullable: true
          description: Vorschaubild-URL des Wikipedia-Artikels, sofern vorhanden.
      required:
        - title
        - extract
        - url
        - lang
...